Reactor code after review
